Use older version of boulder for historic ceremonies and store all output (#13)
The 2000, 2015, and 2020 historical ceremonies used an
[improper](https://github.com/letsencrypt/boulder/commit/ade2bdf564)
version of boulder that would compute the Subject Key Identifier (SKID)
using SHA256 instead of SHA1. Those computation changes are intended and
correct, but were not intended to be used for the historic ceremonies
because their output would not match the real roots and intermediates we
use. This manifested itself as showing a 2023 intermediate signed by the
2020 root-x2 as having its Authority Key Identifier (AKID) be a SHA256
value which is incorrect. @aarongable detected this, thank you.