Update past ceremonies in preparation for next ceremony (#16)
Improve the README so that the process of preparing for future
ceremonies is easier to pick up. Improve the various run.sh scripts so
that their function and purpose is clearer to the reader after several
years of not touching this repo, so that they're more resilient to
changes in softhsm2 (specifically, that it no longer auto-detects the
config file in the root of this repo), and so that each year can point
to an individual ceremony tool version. Finally, update the "2023"
ceremony to accurately reflect the details of the ceremony we actually
conducted in 2024.