hlint: add signature verification
For each certificate, find its issuers (by SKID and by Issuer DN), and verify
each signature on it.

This introduces an error if no issuer can be found for a given certificate,
which means either the inputs need to be a complete hierarchy or the operator
needs to selectively ignore some errors. This was necessary to detect some
problems that existed in a previous version of the `outputs/` directory, where
there was both an Issuer mismatch (due to missing "(FAKE)") and a SKID/AKID
mismatch (due to a change in calculation of SKID).

This also verifies that SKID must be nonempty, and AKID must be nonempty except
for self-signed certificates.