Add hlint tool (#19)
Hwæt! The key-holders in days of old have often held the wrong end of
the stick. Strong chains of scrawled signatures tie anchors to entities,
tie roots of trust to Toots about Rust. But bygone bindings blunder when
bytewise basics are broken. Hlint, eagle-eyed and enterprising, seeks
out SKIDs and Subjects, certifying that same is matched with same.

This checks a pile of certificates for consistent correspondence of
Subject to subjectKeyIdentifier (SKID), and vice versa. If a single
Subject appears with different SKIDs, or a single SKID appears with
different Subjects, it produces an error.

Note that this is an opinionated tool: the is no general prohibition on
distinct Subjects sharing a public key (and thus a SKID), but it is not
our practice or intention to do so.